scripts/validate_data_contract/requirements.txt
databricks-sdk==0.85.0
datacontract-cli[databricks]==0.12.1
